requirements.txt
asgiref==3.2.5
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.14
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-ckeditor==6.7.1
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==9.0.1
pylint==2.4.4
PySocks==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
stripe==2.43.0

